requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
jsonschema==4.20.0
python-multipart==0.0.6
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.1
openai==1.59.7

# Data processing
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Database and storage
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Document processing
beautifulsoup4==4.12.2
lxml==4.9.3
reportlab==4.0.7
Pillow==10.1.0

# Utility libraries
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
jinja2==3.1.2
pyyaml==6.0.1
toml==0.10.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0
